fix price data type for payments, products, and orders

+++ api/internal/product/model.go
@@ -6,14 +6,14 @@ type Product struct {
 	ID          uuid.UUID
 	Name        string
 	Description string
-	Price       float64
+	Price       int64
 }
 
 func (p *Product) GetID() uuid.UUID {
 	return p.ID
 }
 
-func (p *Product) GetPrice() float64 {
+func (p *Product) GetPrice() int64 {
 	return p.Price
 }
